pyproject.toml
[project]
name = "sim-racecenter-agent"
version = "0.1.0"
description = "MCP + Director Agent layer for Sim RaceCenter"
authors = [{name="Sim RaceCenter", email="you@example.com"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.7",
    "anyio>=4.3",
    "asgiref>=3.8.0",
    "click>=8.1",
    "uvicorn>=0.30",
    "fastapi>=0.111",
    "nats-py>=2.6",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "orjson>=3.10",
    "python-dotenv>=1.0",
    "tenacity>=8.2",
    "httpx>=0.27",
    "pypdf>=4.2",
    # Official Model Context Protocol Python SDK (stdio + streamable HTTP transports, tools/resources/prompts framework)
    "mcp[cli]>=0.2.0",
    # Gemini (google) unified Generative AI SDK used for direct MCP tool calling
    "google-genai>=0.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "mypy",
    "ruff",
    "black",
    "isort",
    "pre-commit",
    "types-click",
    "types-orjson"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-q"
markers = [
    "integration: marks tests that require external services like NATS (deselect with -m 'not integration')",
]